The changes below are made to allow two things to happen.  One: sub-detector developers can more readily isolate their own geometry from the global ideal geometry to speed up their development, testing and validation.
Two: The database version can be saved and retrieved in such a way that
the DBIdealGeometryESSource should not retrieve the overlapping information
twice and cause problems with the in-memory graph.  This is most obvious if we
want to use the magnetic field built from MagneticFieldVolumes with and
without the IdealGeometry.  It must work alone and yet not repeat volumes
and placements when used in conjuction with the IdealGeometry.

Since One: allows for each sub-graph, so to speak, to be stored, Two: was
also instrumented as much as possible for sub-detector splitting.